+++ dashboard/modules/serve/tests/test_serve_agent.py
@@ -1,5 +1,4 @@
 import copy
-import subprocess
 import sys
 from typing import Dict
 
@@ -10,38 +9,10 @@ import ray
 from ray import serve
 from ray._private.test_utils import wait_for_condition
 import ray._private.ray_constants as ray_constants
+from ray.tests.conftest import ray_start_stop
 
 GET_OR_PUT_URL = "http://localhost:52365/api/serve/deployments/"
 STATUS_URL = "http://localhost:52365/api/serve/deployments/status"
-
-
-def checkRayStop():
-    try:
-        requests.get("http://localhost:52365/api/ray/version")
-        return False
-    except Exception:
-        return True
-
-
-@pytest.fixture(scope="function")
-def ray_start_stop():
-    subprocess.check_output(["ray", "stop", "--force"])
-    wait_for_condition(
-        lambda: checkRayStop() is True,
-        timeout=15,
-    )
-    subprocess.check_output(["ray", "start", "--head"])
-    wait_for_condition(
-        lambda: requests.get("http://localhost:52365/api/ray/version").status_code
-        == 200,
-        timeout=15,
-    )
-    yield
-    subprocess.check_output(["ray", "stop", "--force"])
-    wait_for_condition(
-        lambda: checkRayStop() is True,
-        timeout=15,
-    )
 
 
 def deploy_and_check_config(config: Dict):
